third_party/blink/web_tests/external/wpt/accname/name/comp_text_node-expected.txt

This is a testharness.js-based test.
Found 5 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] span[role=button] with text node, with leading/trailing non-breaking space
  assert_equals: <span role="button" tabindex="0" class="ex" data-expectedlabel="button&nbsp;label" data-testname="span[role=button] with text node, with leading/trailing non-breaking space">&nbsp;button&nbsp;label&nbsp;</span> expected "button label" but got " button label "
[FAIL] div[role=heading] with text node, with leading/trailing non-breaking space
  assert_equals: <div role="heading" class="ex" data-expectedlabel="heading&nbsp;label" data-testname="div[role=heading] with text node, with leading/trailing non-breaking space">&nbsp;heading&nbsp;label&nbsp;</div> expected "heading label" but got " heading label "
[FAIL] button with text node, with leading/trailing non-breaking space
  assert_equals: <button class="ex" data-expectedlabel="button&nbsp;label" data-testname="button with text node, with leading/trailing non-breaking space">&nbsp;button&nbsp;label&nbsp;</button> expected "button label" but got " button label "
[FAIL] heading with text node, with leading/trailing non-breaking space
  assert_equals: <h3 class="ex" data-expectedlabel="heading&nbsp;label" data-testname="heading with text node, with leading/trailing non-breaking space">&nbsp;heading&nbsp;label&nbsp;</h3> expected "heading label" but got " heading label "
[FAIL] link with text node, with leading/trailing non-breaking space
  assert_equals: <a href="#" class="ex" data-expectedlabel="link&nbsp;label" data-testname="link with text node, with leading/trailing non-breaking space">&nbsp;link&nbsp;label&nbsp;</a> expected "link label" but got " link label "
Harness: the test ran to completion.
